AquaPaintBug.java
This illustrates a possible bug in the Aqua look and feel on the Macintosh. I submitted this bug to Oracle, but they declined to add it to their bug database. I don't know why, but I suspect they concluded the bug was in my code, not theirs. However, as this code can demonstrate, the bug only shows up under the Mac OS X look and feel. This test case lets you try it under any of the four standard looks and feels and it works fine for all except the Mac's AquaLookAndFeel.

The problem is that it this gets a repaint event for every modifiable component in the window, instead of a single repaint event. Normally this wouldn't be a problem, but in this case, I'm using a JLayer to create a blur-effect in the main window when a dialog is open. The blur takes about a third of a second to draw, which is not a serious performance impact. But the JLayer repaints the window once for each paint event. Under Aqua, it sends 15 paint events, so it takes about 5 seconds to draw! You can see this by watching the JCheckBoxes. Aqua initially draws them all enabled, then you can watch Aqua draw them disabled one-by-one while the system is frozen. You also see the count of how many times it repaints in the console output.

Instructions: Run this class. To see the effect, click any of the "Show Message n" buttons. (They all do the same thing.) A dialog will appear, with the blur effect in the main window. Use the floating window next to it to try the effect in different looks and feels. 

On all but the Mac OS X look and feel, it runs fine. On the Mac, it gets a separate repaint event for every component in the window that could be updated. These are the JTextField, the JButtons, and the JCheckBoxes, but not the JLabels. You can see this 2 ways: by watching the buttons as they lose their "selected" state, and by watching the console output, which prints "blurred paint n" each time it gets a repaint message.

I added a lot of buttons are there to amplify the problem. In a real application, there may be many buttons on a tool bar, for example, which would cause the same problem.

Note that unchecking the check boxes doesn't help, even though they no longer have need to update their enabled state. (The buttons that say "shared" may be unchecked in a single operation.) It's merely their presence that generates the extra repaint events.